Update dataset_classification.py (#1599)
The indexing of 1 is incorrect. Given a path, it will think the file name is the first thing in the path. To be consistent with line 174 I simply replaced 1 with -1. But can also use os.path.basename(sample).split(".")[0]

I noticed this when I was using the feat_extract_pytorch.py in action_recognition, and tried to use my custom data paths in the txt file pointed to by VAL_ANNO_PATH in the configuration yaml files.